Separated the reset and refresh functionality
To support this separation, I have added two buttons that reset the views for the chart and the graph independently of each other. Now the refresh button gets the new data in and refreshes the graphs and the visual data, but does not reset the graphs to the default view.